requirements.txt
aiohttp==3.12.15
backports.tarfile==1.2.0
docopt==0.6.2
importlib-metadata==8.0.0
inflect==7.3.1
jaraco-functools==4.3.0
jaraco.collections==5.1.0
ossapi==5.3.2
pip-chill==1.0.3
platformdirs==4.2.2
pydantic==2.11.7
pyinstaller==6.15.0
pyqt6==6.9.1
python-dotenv==1.1.1
tomli==2.0.1
tqdm==4.67.1
yarg==0.1.10
